refactor(meetings): update imports for MediaRequestManager and clean up deprecated code
